# HIVE-25346: cleanTxnToWriteIdTable breaks SNAPSHOT isolation