@deniskuzZ 2021-07-30
Would be great if you could run the HMS benchmark and see if that has affected the commit step performace. See if the index on opType improves the. situation.