@deniskuzZ 2021-08-10
> > Would be great if you could run the HMS benchmark and see if that has affected the commit step performance. See if the index on opType improves the situation.
> 
> There is no benchmark currently for the commitTxn() call.

Could we create one or at least perform similar to https://issues.apache.org/jira/browse/HIVE-23104?focusedCommentId=17083005&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-17083005 test?