@zchovan 2021-09-08
@deniskuzZ I did some benchmarks on commitTxns, please find the results below:


`
Before HIVE-25346

Result "org.apache.hadoop.hive.metastore.tools.ACIDBenchmarks.TestOpenTxn.openTxn":
  N = 100
  mean =     58.136 ±(99.9%) 1.757 ms/op

  Histogram, ms/op:
    [40.000, 42.500) = 0
    [42.500, 45.000) = 0
    [45.000, 47.500) = 0
    [47.500, 50.000) = 2
    [50.000, 52.500) = 11
    [52.500, 55.000) = 14
    [55.000, 57.500) = 20
    [57.500, 60.000) = 26
    [60.000, 62.500) = 13
    [62.500, 65.000) = 7
    [65.000, 67.500) = 1
    [67.500, 70.000) = 3
    [70.000, 72.500) = 0
    [72.500, 75.000) = 1
    [75.000, 77.500) = 1

  Percentiles, ms/op:
      p(0.0000) =     49.111 ms/op
     p(50.0000) =     57.856 ms/op
     p(90.0000) =     63.579 ms/op
     p(95.0000) =     67.901 ms/op
     p(99.0000) =     79.537 ms/op
     p(99.9000) =     79.567 ms/op
     p(99.9900) =     79.567 ms/op
     p(99.9990) =     79.567 ms/op
     p(99.9999) =     79.567 ms/op
    p(100.0000) =     79.567 ms/op


After HIVE-25346


Result "org.apache.hadoop.hive.metastore.tools.ACIDBenchmarks.TestOpenTxn.openTxn":
  N = 100
  mean =     51.015 ±(99.9%) 2.947 ms/op

  Histogram, ms/op:
    [30.000, 35.000) = 0
    [35.000, 40.000) = 3
    [40.000, 45.000) = 26
    [45.000, 50.000) = 27
    [50.000, 55.000) = 17
    [55.000, 60.000) = 11
    [60.000, 65.000) = 9
    [65.000, 70.000) = 4
    [70.000, 75.000) = 1
    [75.000, 80.000) = 1
    [80.000, 85.000) = 1

  Percentiles, ms/op:
      p(0.0000) =     38.385 ms/op
     p(50.0000) =     48.248 ms/op
     p(90.0000) =     64.176 ms/op
     p(95.0000) =     68.280 ms/op
     p(99.0000) =     81.816 ms/op
     p(99.9000) =     81.838 ms/op
     p(99.9900) =     81.838 ms/op
     p(99.9990) =     81.838 ms/op
     p(99.9999) =     81.838 ms/op
    p(100.0000) =     81.838 ms/op

`