@deniskuzZ 2021-09-10
@zchovan thank you for the benchmarks! 
Last thing, to be on a safe side, could you please record INSERTS only if min_history is not available + use old logic for writeId removal based on min_history if available. @pvary, what do you think?